requirements.txt
aniso8601==3.0.2
atomicwrites==1.1.5
attrs==18.1.0
click==6.7
colorama==0.3.9
flasgger==0.9.0
Flask==1.0.2
Flask-JWT-Extended==3.10.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
mistune==0.8.3
more-itertools==4.2.0
nose==1.3.7
pluggy==0.6.0
psycopg2==2.7.5
py==1.5.4
PyJWT==1.6.4
python-dateutil==2.8.0
pytz==2018.4
PyYAML==4.2b1
six==1.11.0
SQLAlchemy==1.2.8
SQLAlchemy-Utils==0.33.11
text-unidecode==1.2
Werkzeug==0.14.1
gunicorn
